Move markers, cleanup re-exports
