+++ src/cli/commands/integrate/git/tools/native/shell-fragments.ts
@@ -19,8 +19,25 @@
  */
 
 import type { GitHookType } from '../../options';
-import { getNativeHookMarker } from '../markers';
-import { resolveSonarHookCommand, SONAR_HOOK_SKIP_SECRETS_MESSAGE } from '../shared';
+import {
+  LEGACY_HOOK_MARKER,
+  resolveSonarHookCommand,
+  SONAR_HOOK_SKIP_SECRETS_MESSAGE,
+} from '../shared';
+
+const NATIVE_HOOK_MARKERS: Record<GitHookType, string> = {
+  'pre-commit': 'sonar pre-commit hook - installed by sonar integrate git',
+  'pre-push': 'sonar pre-push hook - installed by sonar integrate git',
+};
+
+export function getNativeHookMarker(hook: GitHookType): string {
+  return NATIVE_HOOK_MARKERS[hook];
+}
+
+/** New + legacy markers a native hook of this type may legitimately contain (overwrite guard + detection). */
+export function getRecognizedNativeMarkers(hook: GitHookType): string[] {
+  return [NATIVE_HOOK_MARKERS[hook], LEGACY_HOOK_MARKER];
+}
 
 function nativeBinBlock(): string {
   return [
